Refactor useSageAuth into a hook
